feat: Product API 호출 및 adgroup 내 상품 매칭 기능 개발

+++ ad-dsp-api/src/main/java/com/addsp/api/infrastructure/client/dto/ProductApiResponse.java
@@ -9,10 +9,11 @@ import java.time.LocalDateTime;
 public record ProductApiResponse(
         Long id,
         String name,
-        BigDecimal price,
+        Long price,
         String imageUrl,
         Long partnerId,
         LocalDateTime createdAt,
         LocalDateTime modifiedAt
 ) {
 }
+
